+++ torchmetrics/functional/classification/precision_recall.py
@@ -16,23 +16,29 @@ from typing import Optional, Tuple
 import torch
 from torch import Tensor
 
-from torchmetrics.classification.stat_scores import _reduce_stat_scores
-from torchmetrics.functional.classification.stat_scores import _stat_scores_update
+from torchmetrics.functional.classification.stat_scores import _reduce_stat_scores, _stat_scores_update
 from torchmetrics.utilities import _deprecation_warn_arg_is_multiclass, _deprecation_warn_arg_multilabel
+from torchmetrics.utilities.enums import AverageMethod, MDMCAverageMethod
 
 
 def _precision_compute(
     tp: Tensor,
     fp: Tensor,
-    tn: Tensor,
     fn: Tensor,
     average: str,
     mdmc_average: Optional[str],
 ) -> Tensor:
-    # todo: `tn` is unused
+    numerator = tp
+    denominator = tp + fp
+    if average == AverageMethod.NONE and mdmc_average != MDMCAverageMethod.SAMPLEWISE:
+        # a class is not present if there exists no TPs, no FPs, and no FNs
+        meaningless_indeces = torch.nonzero((tp | fn | fp) == 0).cpu()
+        numerator[meaningless_indeces, ...] = -1
+        denominator[meaningless_indeces, ...] = -1
+
     return _reduce_stat_scores(
-        numerator=tp,
-        denominator=tp + fp,
+        numerator=numerator,
+        denominator=denominator,
         weights=None if average != "weighted" else tp + fn,
         average=average,
         mdmc_average=mdmc_average,
@@ -83,6 +89,9 @@ def precision(
 
             .. note:: What is considered a sample in the multi-dimensional multi-class case
                 depends on the value of ``mdmc_average``.
+
+            .. note:: If ``'none'`` and a given class doesn't occur in the `preds` or `target`,
+                the value for the class will be ``nan``.
 
         mdmc_average:
             Defines how averaging is done for multi-dimensional multi-class inputs (on top of the
@@ -178,7 +187,7 @@ def precision(
         raise ValueError(f"The `ignore_index` {ignore_index} is not valid for inputs with {num_classes} classes")
 
     reduce = "macro" if average in ["weighted", "none", None] else average
-    tp, fp, tn, fn = _stat_scores_update(
+    tp, fp, _, fn = _stat_scores_update(
         preds,
         target,
         reduce=reduce,
@@ -190,23 +199,27 @@ def precision(
         ignore_index=ignore_index,
     )
 
-    return _precision_compute(tp, fp, tn, fn, average, mdmc_average)
+    return _precision_compute(tp, fp, fn, average, mdmc_average)
 
 
 def _recall_compute(
     tp: Tensor,
     fp: Tensor,
-    tn: Tensor,
     fn: Tensor,
     average: str,
     mdmc_average: Optional[str],
 ) -> Tensor:
-    # todo: `tp` is unused
-    # todo: `tn` is unused
+    numerator = tp
+    denominator = tp + fn
+    if average == AverageMethod.NONE and mdmc_average != MDMCAverageMethod.SAMPLEWISE:
+        # a class is not present if there exists no TPs, no FPs, and no FNs
+        meaningless_indeces = ((tp | fn | fp) == 0).nonzero().cpu()
+        numerator[meaningless_indeces, ...] = -1
+        denominator[meaningless_indeces, ...] = -1
     return _reduce_stat_scores(
-        numerator=tp,
-        denominator=tp + fn,
-        weights=None if average != "weighted" else tp + fn,
+        numerator=numerator,
+        denominator=denominator,
+        weights=None if average != AverageMethod.WEIGHTED else tp + fn,
         average=average,
         mdmc_average=mdmc_average,
     )
@@ -256,6 +269,9 @@ def recall(
 
             .. note:: What is considered a sample in the multi-dimensional multi-class case
                 depends on the value of ``mdmc_average``.
+
+            .. note:: If ``'none'`` and a given class doesn't occur in the `preds` or `target`,
+                the value for the class will be ``nan``.
 
         mdmc_average:
             Defines how averaging is done for multi-dimensional multi-class inputs (on top of the
@@ -351,7 +367,7 @@ def recall(
         raise ValueError(f"The `ignore_index` {ignore_index} is not valid for inputs with {num_classes} classes")
 
     reduce = "macro" if average in ["weighted", "none", None] else average
-    tp, fp, tn, fn = _stat_scores_update(
+    tp, fp, _, fn = _stat_scores_update(
         preds,
         target,
         reduce=reduce,
@@ -363,7 +379,7 @@ def recall(
         ignore_index=ignore_index,
     )
 
-    return _recall_compute(tp, fp, tn, fn, average, mdmc_average)
+    return _recall_compute(tp, fp, fn, average, mdmc_average)
 
 
 def precision_recall(
@@ -413,6 +429,9 @@ def precision_recall(
 
             .. note:: What is considered a sample in the multi-dimensional multi-class case
                 depends on the value of ``mdmc_average``.
+
+            .. note:: If ``'none'`` and a given class doesn't occur in the `preds` or `target`,
+                the value for the class will be ``nan``.
 
         mdmc_average:
             Defines how averaging is done for multi-dimensional multi-class inputs (on top of the
@@ -509,7 +528,7 @@ def precision_recall(
         raise ValueError(f"The `ignore_index` {ignore_index} is not valid for inputs with {num_classes} classes")
 
     reduce = "macro" if average in ["weighted", "none", None] else average
-    tp, fp, tn, fn = _stat_scores_update(
+    tp, fp, _, fn = _stat_scores_update(
         preds,
         target,
         reduce=reduce,
@@ -521,7 +540,7 @@ def precision_recall(
         ignore_index=ignore_index,
     )
 
-    precision = _precision_compute(tp, fp, tn, fn, average, mdmc_average)
-    recall = _recall_compute(tp, fp, tn, fn, average, mdmc_average)
+    precision_ = _precision_compute(tp, fp, fn, average, mdmc_average)
+    recall_ = _recall_compute(tp, fp, fn, average, mdmc_average)
 
-    return precision, recall
+    return precision_, recall_